backend/requirements.txt
# Web Framework
fastapi>=0.112.0
uvicorn>=0.30.0
pydantic>=2.8.0

# Database
motor==3.3.2           
pymongo==4.6.2

# Machine Learning & Data Science
lightgbm
scikit-learn
pandas
numpy
scipy
joblib

# Computer Vision & OCR
pytesseract
Pillow
torch
torchvision

# Utilities
python-multipart 
python-dotenv

passlib
bcrypt
PyJWT
email-validator

# Rule Engine
pyyaml